requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
libvalkey==4.0.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
requests==2.32.3
setuptools==78.1.0
urllib3==2.3.0
valkey==6.1.0
Werkzeug==3.1.3
wheel==0.45.1
